requirements.txt
pandas==1.5.3
Pillow==9.4.0
scikit_learn==1.2.1
seaborn==0.12.2
